LayoutTests/accessibility/mac/selection-notification-focus-change-expected.txt

This tests that selection changes as a result of advancing focus include AXTextSelectionChangedFocus flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS axTextFocusChangeOne is true
PASS axTextFocusChangeTwo is true
PASS accessibilityController.accessibleElementById("1").isFocusable is true
PASS axTextFocusChangeThree is true
PASS axTextFocusChangeFour is true
PASS successfullyParsed is true

TEST COMPLETE

 1
 2
